requirements.txt
# Core AI and Agent Dependencies
google-genai>=0.3.0
google-adk>=0.1.0
a2a-sdk[http-server]>=0.3.8

# Web Framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
starlette>=0.37.0

# Utility Libraries
python-dotenv>=1.0.0
httpx>=0.27.0
litellm>=1.0.0

# Development and Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
requests>=2.31.0
